replica set working now

+++ index.js
@@ -2,10 +2,10 @@
 
 var _ = require('lodash'),
   debug = require('debug')('mongo-replica-set'),
+  fs = require('fs'),
   EventEmitter2 = require('eventemitter2').EventEmitter2,
   Q = require('bluebird'),
   mongodb = require('mongodb'),
-  MongoClient = Q.promisifyAll(mongodb.MongoClient),
   path = require('path'),
   shell = require('shelljs'),
   util = require('util');
@@ -20,6 +20,7 @@ var _ = require('lodash'),
  * @param {Number} [options.numInstances] no. of instances to create (default is 3).
  * @param {Object} [options.startPort] Port number to start allocating at (default is 27117).
  * @param {Object} [options.baseFolder] Base folder for instance data (default is system temporary folder).
+ * @param {Boolean} [options.logToConsole] Whether to log to console. Default is `false`.
  * @constructor
  */
 var ReplicaSet = exports.ReplicaSet = function(options) {
@@ -31,6 +32,7 @@ var ReplicaSet = exports.ReplicaSet = function(options) {
   this.options = _.extend({
     numInstances: 3,
     startPort: 27117,
+    logToConsole: false,
     baseFolder: path.join(shell.tempdir(), this.name),
   }, options);
 
@@ -45,17 +47,29 @@ util.inherits(ReplicaSet, EventEmitter2);
  * @return {Promise}
  */
 ReplicaSet.prototype.stop = function() {
+  var self = this;
+
   debug('Stopping replica set');
 
-  this.processes.map(function(p) {
-    p.kill();
-  });
+  return Q.map(self.processes, function(p) {
+    // if not yet killed then kill and wait
+    if (!p.killed) {
+      return new Q(function(resolve, reject) {
+        p.on('exit', function() {
+          resolve();
+        });
 
-  shell.rm('-rf', this.options.baseFolder);
+        p.kill();
+      });
+    } else {
+      return true;
+    }
+  })
+    .then(function() {
+      shell.rm('-rf', self.options.baseFolder);
 
-  this.emit('stopped');
-
-  return Q.resolve();
+      self.emit('stopped');
+    });
 };
 
 
@@ -70,7 +84,7 @@ ReplicaSet.prototype.start = function() {
 
   debug('Initialise replica set ' + self.name);
 
-  for (var i=0; i<self.options.numInstances; ++i) {
+  _.times(self.options.numInstances, function(i) {
     var instanceDataFolder = path.join(self.options.baseFolder, 'data' + i);
 
     debug('Create data folder: ', instanceDataFolder);
@@ -79,64 +93,69 @@ ReplicaSet.prototype.start = function() {
 
     debug('Launch instance ' + i);
 
-    self.processes.push(
-      shell.exec(
-        'mongod --port ' + (self.options.startPort + i) + ' --dbpath ' + instanceDataFolder + ' --replSet ' + self.name + ' --smallfiles --oplogSize 128',
-        { 
-          async: true,
-          silent: true,
-        }
-      )
+    var process = shell.exec(
+      'mongod --port ' + (self.options.startPort + i) + ' --dbpath ' + instanceDataFolder + ' --replSet ' + self.name + ' --smallfiles --oplogSize 128',
+      { 
+        async: true,
+        silent: !self.options.logToConsole,
+      }
     );
-  }
 
-  self.emit('instances_launched');
-
-  // connect to first instance to set things up
-  return MongoClient.connectAsync('mongodb://127.0.0.1:' + self.options.startPort)
-    .then(function(db) {
-      debug('Connected to first instance');
-
-      Q.promisifyAll(db);
-
-      return db.evalAsync('rsconf = { _id: "' + self.name + '", members: [] }')
-        .then(function() {
-          return db.evalAsync('rs.initiate(rsconf)');
-        })
-        .then(function() {
-          debug('Initiated replica set');
-
-          // wait a few seconds for replica set to initiate
-          return Q.timeout(3000);
-        })
-        .then(function() {
-          debug('Add remaining instances to set');
-
-          var promises = [];
-          for (var i=1; i<self.options.numInstances; ++i) {
-            promises.push(
-              db.evalAsync('rs.add("127.0.0.1:' + (self.options.startPort + i) + '");')
-            );
-          }
-
-          return Q.all(promises);
-        })
-        .then(function() {
-          debug('Check replica set status');
-
-          return db.evalAsync('rs.status();');
-        })
-        .then(function(data) {
-          debug('Replica set ready: ' + JSON.stringify(data));
-
-          self.emit('ready');
-        });
-    })
-    .catch(function(err) {
-      console.error(err);
-
-      throw err;
+    process.on('exit', function() {
+      process.killed = true;
     });
+
+    self.processes.push(process);    
+  });
+
+  // wait for processes to startup
+  return Q.delay(2000)
+    .then(function() {
+      var killStatus = _.compact(_.pluck(self.processes, 'killed'));
+
+      if (killStatus.length) {
+        throw new Error('Some instances failed to launch');
+      }
+    })
+    .then(function() {
+      self.emit('instances_launched');
+
+      var scriptPath = path.join(self.options.baseFolder, 'setup.js');
+
+      debug('Creating shell script: ' + scriptPath);
+
+      var lines = [
+        'rs.initiate({ _id: "' + self.name + '", members: [ { "_id": 1, "host": "127.0.0.1:' + self.options.startPort + '"} ] });',
+      ]
+      for (var i=1; i<self.options.numInstances; ++i) {
+        lines.push('rs.add("127.0.0.1:' + (self.options.startPort + i) + '");');
+      }
+
+      fs.writeFileSync(scriptPath, lines.join('\n'));
+
+      debug('Executing shell script');
+
+      self._exec('mongo --port ' + self.options.startPort + ' ' + scriptPath, 'setup-output.txt');
+
+      debug('Checking replica status');
+
+      self._exec('mongo --port ' + self.options.startPort + ' --eval "JSON.stringify(rs.status());"', 'status-output.txt');
+
+      debug('Replica set ready');
+
+      self.emit('ready');
+    });
+};
+
+
+
+ReplicaSet.prototype._exec = function(command, outputFileName) {
+  var suffix = '';
+  if (!this.options.logToConsole) {
+    var suffix = ' >> ' + path.join(this.options.baseFolder, outputFileName);
+  } 
+
+  shell.exec(command + suffix);
 };
